# [not for merging] SQlite Base Node Indices