--- a/.gitignore
+++ b/.gitignore
@@ -4,6 +4,7 @@ run.sh
 *-log-*.json
 *.png
 *.log
+soup.db
 plotting/*.png
 
 callgrind.out.*