--- a/core/src/lib.rs
+++ b/core/src/lib.rs
@@ -12,9 +12,11 @@ extern crate nom_sql;
 extern crate petgraph;
 extern crate rahashmap;
 extern crate rand;
+extern crate rusqlite;
 extern crate serde;
 #[macro_use]
 extern crate serde_derive;
+extern crate serde_json;
 
 pub mod addressing;
 pub mod data;