--- a/core/src/lib.rs
+++ b/core/src/lib.rs
@@ -25,7 +25,7 @@ pub mod map;
 
 pub use data::{DataType, Datas, Record, Records};
 pub use addressing::{IndexPair, LocalNodeIndex};
-pub use local::{KeyType, LookupResult, MemoryState, Row, State, Tag};
+pub use local::{KeyType, LookupResult, Row, State, Tag};
 pub use map::Map;
 pub use petgraph::graph::NodeIndex;