--- a/core/src/local/mod.rs
+++ b/core/src/local/mod.rs
@@ -6,7 +6,7 @@ mod state;
 mod keyed_state;
 
 pub use data::{DataType, SizeOf};
-pub use self::state::State;
+pub use self::state::{MemoryState, State};
 
 #[derive(Clone, Copy, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Serialize, Deserialize)]
 pub struct Tag(pub u32);