--- a/core/src/local/state.rs
+++ b/core/src/local/state.rs
@@ -6,8 +6,8 @@ use std::rc::Rc;
 use ::*;
 use data::SizeOf;
 use local::single_state::SingleState;
-use serde_json;
 
+use bincode;
 use rand::{self, Rng};
 use rusqlite::{self, Connection};
 use rusqlite::types::{ToSql, ToSqlOutput};
@@ -190,8 +190,8 @@ impl PersistentState {
 
     // Used with statement.query_map to deserialize the rows returned from SQlite
     fn map_rows(result: &rusqlite::Row) -> Vec<DataType> {
-        let row: String = result.get(0);
-        serde_json::from_str(&row).unwrap()
+        let row: Vec<u8> = result.get(0);
+        bincode::deserialize(&row).unwrap()
     }
 
     fn add_key(&mut self, columns: &[usize], partial: Option<Vec<Tag>>) {
@@ -231,7 +231,7 @@ impl PersistentState {
         // Make sure that existing rows contain the new index column too,
         // if there are any. This could be faster with UPDATE statements
         // if needed in the future.
-        if self.len() > 0 {
+        if self.rows() > 0 {
             let rows = self.cloned_records();
             self.clear();
             for row in rows {
@@ -268,7 +268,7 @@ impl PersistentState {
             ))
             .unwrap();
 
-        let row = serde_json::to_string(&r).unwrap();
+        let row = bincode::serialize(&r).unwrap();
         let mut values: Vec<&ToSql> = vec![&row];
         let mut index_values = self.indices
             .iter()