--- a/core/src/local/state.rs
+++ b/core/src/local/state.rs
@@ -176,6 +176,7 @@ impl PersistentState {
         connection
             .execute_batch(
                 "CREATE TABLE IF NOT EXISTS store (row BLOB);
+                PRAGMA locking_mode = EXCLUSIVE;
                 PRAGMA synchronous = NORMAL;
                 PRAGMA journal_mode = WAL;",
             )