… Cargo.lock
… benchmarks/vote/clients/localsoup/mod.rs
… benchmarks/vote/main.rs
… core/Cargo.toml
… dataflow/src/domain/mod.rs
… dataflow/src/lib.rs
… dataflow/src/node/process.rs
… dataflow/src/ops/filter.rs
… dataflow/src/ops/grouped/mod.rs
… dataflow/src/ops/join.rs
… dataflow/src/ops/latest.rs
… dataflow/src/ops/mod.rs
… dataflow/src/ops/topk.rs
… dataflow/src/persistence/mod.rs
… dataflow/src/processing.rs
… examples/basic-recipe.rs
… src/bin/souplet.rs
… src/lib.rs
… src/tests.rs