This implements persistent indices for Soup base nodes using SQlite, as a baseline for future experiments.

### Overview
At an API level it converts `State` into an enum that can be either `MemoryState` or `PersistentState` - with the latter now being used for base nodes. 

### Storage
Each instance of `PersistentState` corresponds to one table in SQlite. The table contains one column for each index and a single `row` column that stores a serialized representation of the `Vec<DataType>`. This might not be the right approach, so let me know if you have any better ideas!

#### Adding indices
`add_key` makes sure that there's a separate `index_#` column for all the parts of the given `columns` and finally adds a combined index using `CREATE INDEX`. This means that any rows that were inserted prior to `add_key` being called won't contain the new index columns though, so might be worth going through and backfilling those.

#### Look-ups
`lookup` builds up a `SELECT` query using the indices available in `columns`, on the form of `SELECT row FROM table_name WHERE index_0 = x AND index_1 = x AND ....`.

#### Removals
`remove` plucks out all the index values from the given row and performs a `DELETE FROM` with those.

#### Inserts
`insert` finds all the index values from the given row (from `self.indices`) and inserts them into separate `index_#` columns. The whole row gets put into a `row` column, after being serialized to JSON. This could probably be a `BLOB` column and use `bincode` instead as well.

### Future work
This mostly just implements the API - haven't run any benchmarks yet. This also performs an `INSERT` on every batch update, which with autocommit probably causes a disk write. @jonhoo suggested earlier that a possible solution might write on every update, but only flush to disk before the batch packet is inserted into the graph. Maybe this could be implemented by creating a transaction, and only committing it prior to insertion into the graph - not sure yet.