@jonhoo 2018-02-16
Maybe I'm misremembering, but I think lookups should only ever use exactly one index. I don't think you should ever end up having to `AND` them together. You'd have to look at the `add_key`  for the old in-memory `State` to see whether it expects old rows to also be automatically be added to new indices. It probably does if something is a full materialization.

Some things to be aware of if you want this to also work for partial materalizations.
 - When a new index is added, it may be that existing rows do not need to be added to that index eagerly. In fact, I'm fairly sure that's not the case. But you should check the code.
 - Be aware that (for partial) you *need* to store empty results for keys too! That is, you need to be able to distinguish between `lookup(3)` returning `Missing` and returning no rows. The former requires a replay, the latter does not.