@ekmartin 2018-02-18
@jonhoo: probably described it poorly - what I meant to say was that lookups `AND` together the columns of the given index, i.e. the `columns` argument given to `lookup`. That way it works for both single column and combined indices. I added a test with lookups against different indices as well [here](https://github.com/ekmartin/distributary/blob/95db92006ef59f490a6e3f3c2940c8cb37d3b2ee/core/src/local.rs#L1439-L1465).

I added a small fix that makes sure existing rows work with new indices as well.

Haven't looked into making it work with partial indices yet, but those are good points if I eventually need this to work with other nodes than bases.